Add reading/writing of pe/coff debug directory and codeview records
Add routines for reading and writing of the debug directory entries in a pe/coff
header.

Also add routines for reading and writing a codeview record which may be pointed
to by a debug directory entry.

This tries to match the existing style, but not sure if it succeeds. Not sure
how to handle the variable length pdbfilename the codeview record contains, but
it's not very interesting, so just leave that out for the moment.

bfd/ChangeLog:

2014-01-20  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* peXXigen.c (_bfd_XXi_swap_debugdir_in)
	(_bfd_XXi_swap_debugdir_out, _bfd_XXi_slurp_codeview_record)
	(_bfd_XXi_write_codeview_record): Add functions for reading and
	writing debugdir and codeview records.
	* libpei.h (_bfd_XXi_swap_debugdir_in)
	(_bfd_XXi_swap_debugdir_out, _bfd_XXi_write_codeview_record):
	Add prototypes and macros.

include/coff/ChangeLog:

2014-01-20  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* pe.h (external_IMAGE_DEBUG_DIRECTORY, _CV_INFO_PDB70)
	(_CV_INFO_PDB20): Add structures and constants for debug directory
	and codeview records.
	* internal.h (internal_IMAGE_DEBUG_DIRECTORY, CODEVIEW_INFO):
	Add structures and constants for internal representation of debug
	directory and codeview records.