Factor out common code for supporting 'ld --buildid'
Factor out the code for parsing the 'ld --buildid' option and calculating the
build-id, which will be common between ELF and PE/COFF implementations, as
ldbuildid.c.  Place prototypes in ldbuildid.h.

uuid-style buildid can't be supported on MinGW due to lack of /dev/urandom.

XXX: This doesn't include the regeneration of ld/Makefile.in, as I don't have
the correct automake version available.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>

ld/ChangeLog:

2014-01-20  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* emultempl/elf32.em (id_note_section_size, read_hex, write_build_id):
	Move code for parsing build-id option and calculating the build-id to...
	* ldbuildid.c: New file.
	* ldbuildid.h: New file.
	* Makefile.am (CFILES, HFILES, OFILES, ld_new_SOURCES): Add new files.