Add ld --build-id support for pe/coff
Add ld --build-id option for pe/coff, which adds a .build-id section containing
a CodeView record, pointed to by the debug directory.

Note that the pe/coff build-id is a fixed length of 16 bytes, not a variable
length like an ELF .note.gnu.build-id.

XXX: There is an unfortunate degree of code duplication between 32 and 64 bit
pe/coff variants, but this is hard to avoid given the way the code is
structured. :-(

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>

bfd/ChangeLog:
2014-01-20  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* libcoff-in.h (pe_tdata): Add build-id data.
	* libcoff.h: Regenerate.
	* coffcode.h (coff_write_object_contents): Run build_id
	after_write_object_contents hook.

ld/ChangeLog:

2014-01-20  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* emultempl/pe.em (gld${EMULATION_NAME}_handle_option):
	(pecoff_checksum_contents, write_build_id, setup_build_id)
	(gld_${EMULATION_NAME}_after_open):  Handle and implement
	build-id option.
	* emultempl/pep.em: Ditto.