pe/coff: Don't break .build-id with objcopy/strip
The file offsets in the debug directory need to be recalculated in
_bfd_XX_bfd_copy_private_bfd_data_common(), so that if an objcopy or strip adds
or removes something which causes the file offset of the CV record to change,
those offsets are correct.

(This can happen due to either sections being added or removed before the
.build-id section directly affecting it's file offset, or by the addition or
removal of sufficent section headers to change the first section's file offset
and hence all subsequent sections' file offsets.)

(In general _bfd_XX_bfd_copy_private_bfd_data_common() could be a whole lot
smarter about adjusting the DataDirectory.  If a section corresponding to a RVA
is removed or no longer has any contents, that entry should be cleared?  Offsets
should always be recomputed?)

bfd/ChangeLog:

2014-04-10  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* peXXigen.c (is_vma_in_section, find_section_by_vma): New
	functions.
	(_bfd_XX_bfd_copy_private_bfd_data_common): Recalculate file
	offsets in the debug directory.