7.8-windows-nat-cygwin.patch
Remaining parts of 7.8-windows-nat-cygwin.patch not yet upstream.

* Replace have_saved_context with signal_thread_id throughout, and store
the thread id in it

* In thread_rec(), only use the signal saved context if we are retrieving the
context for the correct thread.

* Clear ContextFlags in the signal saved context so we never attempt to restore
it to the inferior

Since the signal context is the context which will be restored after any signal
handler has been run, to resume to it skips over the actual signal delivery.

(Unfortunately, this isn't quite right.  If GDB decides it needs to single-step
after continuing by setting FLAG_TRACE_BIT, we must alter the inferior state.
So we potentially need to have both the signal saved context, and the actual
context for the thread.  Which is not a straightforward change to make.)

Unrelated changes:

* Ignore ERROR_INVALID_HANDLE from SuspendThread()

I have no testcase to demonstrate this. Discussed somewhat in the thread
https://cygwin.com/ml/gdb-patches/2013-06/msg00680.html

* Remove __COPY_CONTEXT_SIZE conditional compilation.  These are all under
__CYGWIN__ already, and versions of Cygwin which don't define
__COPY_CONTEXT_SIZE are long obsolete.

* Trim any trailing '\n's from OutputDebugString before echoing it.
(Also an unclear change for detecting if it's a cygwin message, but checking
for a leading 'cYg' should be enough)

* Cast DWORD thread_id to int before returning it from get_windows_event.
(I don't think this is needed.  If DWORD->int is a narrowing conversion, we
have other problems...)

* In one place, clarify that thread_rec()'s second parameter is a bool flag.
(This should be done everywhere, or not)