Teach gdb how to unwind cygwin _sigbe frames
I really wanted to do this by adding some DWARF CFI to the generated sigfe.s
file, but there doesn't seem to currently be a way to correctly describe
_sigbe's frame using that.

So instead, write a custom unwinder for _sigbe frames, which gets the return
address from the sigstack.

Only implemented for amd64 for the moment, although i386 should be mechanically
very similar.

Testcases:

Set a breakpoint on a _sigfe wrapped function when running a program which calls
it

With 'set cygwin-exceptions on', try the sample from:
https://www.sourceware.org/ml/gdb/2013-10/msg00167.html

Issues:

Should detect if we are in _sigbe after the return address has been popped of
the sigstack, and if so, fetch the return address from the register instead.

If a signal has actually arrived while inside the sigfe-wrapped function, we
will return via sigdelayed which has been pushed onto the sigstack, rather than
directly to the caller.  sigdelayed's stack frame probably also needs special
handling if we want to unwind it correctly.  I'm not sure how to test that case.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>