Teach gdb how to unwind cygwin _sigbe frames
I really wanted to do this by adding some DWARF CFI to the generated sigfe.s
file, but there doesn't seem to currently be a way to correctly describe
_sigbe's frame using that.

So instead, write a custom unwinder for _sigbe frames, which gets the return
address from the sigstack.

Implemented for i386 and amd64.

Testcases:

1. Set a breakpoint on a _sigfe wrapped function in a program which calls it.
When the breakpoint is hit, examine the backtrace.

2. With 'set cygwin-exceptions on', try the sample from:
https://www.sourceware.org/ml/gdb/2013-10/msg00167.html
When the exception is hit, examine the backtrace.

Issues:

1. This should detect if we are in _sigbe after the return address has been
popped off the sigstack, and if so, fetch the return address from the register
it's been popped into instead.

2. If a signal has actually arrived while inside the sigfe-wrapped function, we
will return via sigdelayed which has been pushed onto the sigstack, rather than
directly to the caller.

We probably need to give sigdelayed's stack frame the same special handling, so
that we can backtrace from the signal handler correctly.

3. If there are multiple sigbe or sigdelayed stack frames to be unwound, this
only unwinds the first one correctly, because we don't unwind the value of
sigstackptr itself.

This is no worse than currently, when we can't even unwind one sigbe frame
correctly, but isn't quite correct.

Presumably that can happen if a signal handler calls a cygwin DLL function which
is wrapped by _sigfe.

If we care about that, I guess we would need to define a pseduo-register to
track it's value as we unwind the stack.

4. This unfortunately ends up hardcoding into gdb the offset of sigstackptr in
the cygwin TLS area from the top of stack, and so will break if that changes.
Hopefully that doesn't happen to often.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>