Don't set breakpoints on import stubs on Windows amd64
On Windows amd64, setting a breakpoint on a symbol imported from a shared
library sets the breakpoint on both the import stub, and in the shared library.

This is due to the minimal symbol for the import stub not being correctly given
the type mst_solib_trampoline (unlike Windows i386).

This looks like an error in commit 303c5, possibly not noticed at the time as
tests were only run on i386.

I'm not sure this is quite right on i386.  It assumes that name[5] == '_', that
is, it assumes that the imported symbol starts with an underscore which is being
silently removed...