bfd: Don't compress a section if it doesn't get smaller
bug #18087

The section virtual size is not changed when a section is compressed or
decompressed.

Normally this is not a problem.

However, when the compressor can't help and produces a larger output section,
disaster strikes, as the data in the section is trucated to the original virtual
size.

Updating the virtual size when the section is compressed doesn't seem possible,
as that breaks the COFF requirement that sections much be contiguous

So, don't compress debug sections if they don't get any smaller.