Fixes in win32pstatus ELF note handling
In elfcore_grok_win32pstatus():

- Don't ignore all notes smaller than minimum size for a NOTE_INFO_THREAD

This only happens to work because the Cygwin dumper tool currently
writes all these notes as the largest size of the union, (which wastes
lots of space in the "core dump").

- Read tid from correct offset in NOTE_INFO_THREAD

This probably meant that registers were only being correctly recovered from
the "core dump" for the current thread.